further clean up shared api
