+++ specs/tui-toolcalling/TECH.md
@@ -45,6 +45,8 @@ pub(crate) trait AgentToolScheduleHost: Sized {
 pub(crate) struct AgentToolScheduler; // generic static methods over the host
 ```
 `AgentToolScheduler` owns `queue_actions`, `try_to_execute_available_actions`, `start_pending_action_by_id`, and `finish_action`. Surface-specific execution (shell, file edits) and side effects (history status, event emission, TUI cards) are delegated to the host. GUI-specific inline views, shared-session viewer UI, terminal block events, `TerminalModel`, `AIBlock`, `RequestedCommandView`, and `CodeDiffView` stay in the GUI adapter.
+Generic state queries also live on `AgentToolActionModel`, not on the surface wrappers. `get_pending_actions`, `get_pending_actions_for_conversation`, `get_pending_action_by_id`, `has_unfinished_actions_for_conversation`, `get_finished_action_results`, `get_action_result`, `restore_action_results_from_exchanges`, `blocked_action_for_conversation`, and `get_action_status` are implemented on the shared model. `get_action_status` takes `is_view_only: bool` as a parameter so the GUI can pass its view-only flag and the TUI can pass `false`; the `Blocked` vs `Queued` distinction is the only status rule that depends on surface state. The surface wrappers keep thin delegating methods so existing GUI view code (`block.rs`, `output.rs`, etc.) continues to call `action_model.get_action_status(id)` unchanged. Running-action recording is consolidated into `AgentToolActionModel::record_running_action` (with the phase-consistency `debug_assert`), so the GUI no longer carries a duplicate `add_running_action`.
+Queries that genuinely depend on surface state stay on the wrappers: `get_pending_action(app)`, `get_pending_or_running_action_id(app)`, `has_unfinished_actions(app)`, and `get_async_running_action(app)` all resolve the active conversation through GUI-specific `terminal_view_id` + `BlocklistAIHistoryModel`, and `get_async_running_action` additionally consults the GUI executor's in-flight action map. `mark_action_as_remotely_executing` stays GUI-specific because it is gated on `is_view_only` and emits a GUI event.
 The two surfaces keep different async-completion mechanisms behind `try_execute`: GUI execution completes via its executor's `FinishedAction` event subscription, TUI execution via its own `ctx.spawn` callback; both call back into `AgentToolScheduler::finish_action`. For v0, TUI auto-accepts (no user-confirmation blocked state) and inherits the shared phased loop, so read-only tools fan out in one parallel phase while shell and file-edit tools run as serial barriers.
 ### Shared-first tool executor
 Tool execution is centralized in a shared `AgentToolExecutor` used by both GUI and TUI. Shared tools are handled directly; inherently surface-specific tools are delegated to a required `SurfaceSpecificToolExecutor` implemented by each surface. `AgentToolExecutor` is a unit type with static methods generic over the surface — it holds no state.
